+++ app/src/androidTest/java/com/github/sdpteam15/polyevents/view/activity/admin/AdminHomeFragmentTest.kt
@@ -14,20 +14,21 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.sdpteam15.polyevents.HelperTestFunction
 import com.github.sdpteam15.polyevents.R
 import com.github.sdpteam15.polyevents.fakedatabase.FakeDatabase
+import com.github.sdpteam15.polyevents.fakedatabase.FakeDatabaseItem
 import com.github.sdpteam15.polyevents.model.database.remote.Database
 import com.github.sdpteam15.polyevents.model.database.remote.FirestoreDatabaseProvider
 import com.github.sdpteam15.polyevents.model.database.remote.login.UserLogin
+import com.github.sdpteam15.polyevents.model.entity.Item
 import com.github.sdpteam15.polyevents.model.entity.UserEntity
 import com.github.sdpteam15.polyevents.model.entity.UserProfile
-import com.github.sdpteam15.polyevents.model.entity.UserRole
 import com.github.sdpteam15.polyevents.model.observable.Observable
-import com.github.sdpteam15.polyevents.model.observable.ObservableList
+import com.github.sdpteam15.polyevents.view.activity.ItemRequestActivity
+import com.github.sdpteam15.polyevents.view.activity.ItemsAdminActivity
 import com.github.sdpteam15.polyevents.view.activity.MainActivity
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.mockito.kotlin.anyOrNull
 import org.mockito.Mockito.`when` as When
 
 @RunWith(AndroidJUnit4::class)
@@ -99,4 +100,37 @@ class AdminHomeFragmentTest {
         Espresso.onView(ViewMatchers.withId(R.id.btnRedirectUserManagement)).perform(click())
         Intents.intended(IntentMatchers.hasComponent(UserManagementListActivity::class.java.name))
     }
+
+    /*
+    @Test
+    fun itemRequestActivityOpensOnClick() {
+        Espresso.onView(ViewMatchers.withId(R.id.ic_more)).perform(click())
+
+        Espresso.onView(ViewMatchers.withId(R.id.btnRedirectItemReqManagement)).perform(click())
+        Intents.intended(IntentMatchers.hasComponent(ItemRequestActivity::class.java.name))
+    }
+
+    @Test
+    fun itemsAdminActivity() {
+        var availableItems: MutableMap<Item, Int> = mutableMapOf()
+        availableItems[Item(null, "Chocolat", "OTHER")] = 30
+        availableItems[Item(null, "Kiwis", "OTHER")] = 10
+        availableItems[Item(null, "230V Plugs", "PLUG")] = 30
+        availableItems[Item(null, "Fridge (large)", "OTHER")] = 5
+        availableItems[Item(null, "Cord rewinder (15m)", "PLUG")] = 30
+        availableItems[Item(null, "Cord rewinder (50m)", "PLUG")] = 10
+        availableItems[Item(null, "Cord rewinder (25m)", "PLUG")] = 20
+
+        Database.currentDatabase = FakeDatabase
+
+        FakeDatabaseItem.items.clear()
+        for ((item, count) in availableItems) {
+            Database.currentDatabase.itemDatabase!!.createItem(item, count)
+        }
+
+        Espresso.onView(ViewMatchers.withId(R.id.ic_more)).perform(click())
+        Espresso.onView(ViewMatchers.withId(R.id.btn_admin_items_list)).perform(click())
+        Intents.intended(IntentMatchers.hasComponent(ItemsAdminActivity::class.java.name))
+    }
+    */
 }
