+++ app/src/main/java/com/github/sdpteam15/polyevents/model/entity/RouteNode.kt
@@ -2,7 +2,6 @@ package com.github.sdpteam15.polyevents.model.entity
 
 import com.github.sdpteam15.polyevents.model.map.Attachable
 import com.github.sdpteam15.polyevents.model.map.LatLngOperator.euclideanDistance
-import com.github.sdpteam15.polyevents.model.map.LatLngOperator.isOnSegment
 import com.google.android.gms.maps.model.LatLng
 
 /**
