+++ app/src/main/java/com/github/sdpteam15/polyevents/model/map/RouteMapHelper.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import android.graphics.Color
 import androidx.lifecycle.LifecycleOwner
 import com.github.sdpteam15.polyevents.R
-import com.github.sdpteam15.polyevents.helper.HelperFunctions
 import com.github.sdpteam15.polyevents.model.database.remote.Database
 import com.github.sdpteam15.polyevents.model.entity.RouteEdge
 import com.github.sdpteam15.polyevents.model.entity.RouteNode
