+++ app/src/main/java/com/github/sdpteam15/polyevents/view/fragments/MoreFragment.kt
@@ -1,15 +1,11 @@
 package com.github.sdpteam15.polyevents.view.fragments
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Button
 import androidx.fragment.app.Fragment
 import com.github.sdpteam15.polyevents.R
-import com.github.sdpteam15.polyevents.view.activity.ItemRequestActivity
-import com.github.sdpteam15.polyevents.view.activity.ItemsAdminActivity
 
 
 /**
