+++ app/src/main/java/com/github/sdpteam15/polyevents/view/fragments/home/ProviderHomeFragment.kt
@@ -2,15 +2,14 @@ package com.github.sdpteam15.polyevents.view.fragments.home
 
 import android.content.Intent
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
+import androidx.fragment.app.Fragment
 import com.github.sdpteam15.polyevents.R
 import com.github.sdpteam15.polyevents.model.entity.UserRole
 import com.github.sdpteam15.polyevents.view.activity.ItemRequestActivity
-import com.github.sdpteam15.polyevents.view.activity.ItemsAdminActivity
 import com.github.sdpteam15.polyevents.view.activity.MainActivity
 
 class ProviderHomeFragment : Fragment() {
